Major refactoring of inference module - Created a InferenceModel class that holds all of the metadata required   to run models from a TrainingJob. This includes input/output scale,   skeletons, Keras model loading and caching, variable input shape,   normalization and other convenience properties. - Added some generic attributes to the sleap.nn.Model class to allow   for storing or inferring output scaling (depends on architecture   details, so it can be overloaded in backbones). - sleap.nn.loadmodel module removed. - Predictor refactored to take advantage of InferenceModel caching and   reuse multiprocessing pool (fixes: #155).   - Previously, models were being reloaded from disk and copies made     every time a new crop size was encountered.   - Previously, a new pool was being opened on every predict() call. - Refactored DataTransform usage to account for discrepancies between   input scale vs output scale vs relative output scale vs scale across   models. - Fix: DataTransform-based scaling would always convert images to   float64 dtype. - Refactored TensorFlow-based peak finding for GPU execution.   - Account for unspecified confmaps tensor at graph creation time.   - Fix subpixel refinement algorithm to enable greater resolution than     original data scale.   - General rework of offset adjustment.   - Exposed supersampling parameters (window size, window batching)   - Fix tensors being overwritten or dropping out of context   - Fixes: #149 - Exposed peak finding options and other inference functionality as   Predictor attributes. - Renamed: Instance.visible_points_array -> Instance.points_array - Sensible cropping defaults in profiles and function kwargs. - Indented default profiles for readability. - General unfucking of syntax errors, anti-patterns, typos, etc.

+++ sleap/nn/transform.py
@@ -79,7 +79,7 @@ class DataTransform:
         if (img_h, img_w) != target_size:
 
             # build ndarray for new size
-            scaled_imgs = np.zeros((imgs.shape[0], h, w, imgs.shape[3]))
+            scaled_imgs = np.zeros((imgs.shape[0], h, w, imgs.shape[3]), dtype=imgs.dtype)
 
             for i in range(imgs.shape[0]):
                 # resize using cv2
